TSK-63: Refactored SummayService/Controller/Test into Task - TaskSummary. REST-Path for TaskSummary updated.
