--- lib/taskana-core/src/main/java/pro/taskana/SummaryService.java
@@ -1,14 +1,0 @@
-package pro.taskana;
-
-import java.util.List;
-
-import pro.taskana.exceptions.WorkbasketNotFoundException;
-import pro.taskana.model.TaskSummary;
-
-/**
- * This interface manages the summaries of some models.
- */
-public interface SummaryService {
-
-    List<TaskSummary> getTaskSummariesByWorkbasketId(String workbasketId) throws WorkbasketNotFoundException;
-}
