+++ lib/taskana-core/src/main/java/pro/taskana/TaskanaEngine.java
@@ -26,13 +26,6 @@ public interface TaskanaEngine {
     ClassificationService getClassificationService();
 
     /**
-     * Can be used for each operation for summaries with short-infos about
-     * one or more objects.
-     * @return a SummaryService-BEAN
-     */
-    SummaryService getSummaryService();
-
-    /**
      * The Taskana configuration.
      * @return the TaskanaConfiguration
      */
